Upgrade system, migrate cobbler data from old container to new
During cluster deployment, orchestrator registers nodes
in cobbler, cobbler keeps all this information in file
system based database which is in /var/lib/cobbler/config
directory, we have to migrate this directory from one
container to another.

We have the next data migartion flow for cobbler
* take /var/lib/cobbler/config from previous container
* save it in working directory in
  /var/lib/fuel_upgrade/version/cobbler_configs
* mount /var/lib/fuel_upgrade/version in container
  as /tmp/upgrade/
* right after cobbler container is created, copy
  cobbler data from mounted directory to
  /var/lib/cobbler/config

blueprint upgrade-to-5-1
Closes-bug: #1331975

+++ fuel_upgrade_system/fuel_upgrade/fuel_upgrade/utils.py
@@ -249,6 +249,10 @@ def copy(from_path, to_path):
 
 def copytree(source, destination, overwrite=True):
     """Copy a given source directory to destination folder.
+
+    :param str source: copy from
+    :param str destination: copy to
+    :param bool overwrite: overwrite destination directory
     """
     logger.debug(u'Copy folder from %s to %s', source, destination)
 
@@ -258,5 +262,25 @@ def copytree(source, destination, overwrite=True):
 
 
 def rmtree(source, ignore_errors=True):
+    """Remove directory
+
+    :param str source: path to directory
+    :param bool ignore_errors: ignores error if True
+    """
     logger.debug(u'Removing %s', source)
-    shutil.rmtree(source, ignore_errors=ignore_errors)
+    if os.path.exists(source):
+        shutil.rmtree(source, ignore_errors=ignore_errors)
+
+
+def check_file_is_valid_json(path):
+    """Checks if file contains valid json
+
+    :param str path: path to json file
+    :returns: True if valid False if invalid
+    """
+    try:
+        json.load(open(path, 'r'))
+    except (ValueError, IOError):
+        return False
+
+    return True
